[VK-3] implemented request for fetching users

+++ VK_Client/app/src/main/java/com/nethergrim/vk/web/WebRequestManager.java
@@ -14,5 +14,5 @@ public interface WebRequestManager {
 
     public void getConversations(int limit, int offset, boolean onlyUnread, int previewLenght, final WebCallback<ConversationsList> callback);
 
-    public void getUsers(List<Integer> ids, List<String> fields, String nameCase, WebCallback<ListOfUsers> callback);
+    public void getUsers(List<Long> ids, List<String> fields, String nameCase, WebCallback<ListOfUsers> callback);
 }
